// frontend/src/types/pagination-response.ts
export type paginationResponseType<T = unknown> = {
  data: T
  current_page: number
  first_page_url: string
  from: number | null
  last_page: number
  last_page_url: string
  next_page_url: string | null
  path: string
  per_page: number
  prev_page_url: string | null
  to: number | null
  total: number
  links: {
    url: string | null
    label: string
    active: boolean
  }[]
}